Change return type of getAllExtensions.
Used to be std::set<std::string>. The set would contain strings
representing the extensions. However, path::extension returns a path
and filesystem::create_directory expects a path. So, if I stored the
extensions as strings, I would have needed to convert them to strings
and then convert them back to paths. So instead, getAllExtensions
is going to return a set of paths since that's what I'm going to need
anyway. For clearity, getAllExtensions now returns a
std::set<std::filesystem::path>

In addition, the function now verifies that the file whose extension
its grabbing has an extension so that we're not grabbing empty paths.